rename index to ping
